riot_buddy/requirements.txt
boto3==1.26.11
email_validator==1.3.0
Flask==2.2.2
Flask_Cors==3.0.10
Flask_Login==0.6.2
flask_sqlalchemy==3.0.2
nanoid==2.0.0
python-dotenv==0.21.0
setuptools==57.4.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
psycopg2-binary==2.9.5
Faker==15.3.2
nanoid==2.0.0
sendgrid==6.9.7
twilio==7.16.0
